LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/interesttarget.tentative-expected.txt

Toggle the popover  Toggle the popover  Toggle the popover

FAIL Shadow root reference target works with interesttarget attribute. assert_equals: showCount expected 1 but got 0
FAIL Shadow root reference target works with interesttarget attribute via options. assert_equals: showCount expected 1 but got 0
FAIL Shadow root reference target works with .interestTargetElement property. assert_equals: showCount expected 1 but got 0
